+++ scripts/rebuild_specs.js
@@ -81,7 +81,7 @@ const conversions = [
  */
 function rebuildConverterTests(dirs) {
     const program = ts.createProgram(app.options.getFileNames(), {
-        ...app.options.getCompilerOptions(),
+        ...app.options.getCompilerOptions(app.logger),
         noEmit: true,
     });